requirements.txt
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.10.0
asgiref==3.9.1
astor==0.8.1
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
Django==4.2.9
django-admin-soft-dashboard==1.0.22
django-cors-headers==4.3.1
django-dbbackup==4.2.1
django-debug-toolbar==4.4.6
django-filter==23.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
filelock==3.19.1
fsspec==2025.7.0
graphviz==0.20.3
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
jiter==0.10.0
numpy==2.3.2
packaging==25.0
pandas==2.2.3
Pillow==10.1.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
str2bool==1.1
tokenizers==0.22.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.7.0
